ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
elasticsearch==8.11.0
sentence-transformers==2.2.2
huggingface-hub==0.16.4
pytesseract==0.3.10
Pillow==10.1.0
requests==2.31.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
numpy==1.26.2
minio==7.2.0
pdf2image==1.16.3
